Production fix
The program used to crash when exposed to a device that did not have any records associated with it. Unfortunately I fixed this last week and cannot recall whether I rectified that problem or simply added a record to the DB. I'm inclined to believe the former after looking at the current changes, but still.